Add recoil state management
