+++ src/App.jsx
@@ -8,10 +8,10 @@ import ContactUs from "./pages/ContactUs";
 import WhyPocketmePage from "./pages/WhyPocketmePage";
 import Loader from "./components/Loader";
 import { useEffect, useState } from "react";
+import { RecoilRoot } from "recoil";
 
 function App() {
   const [loading,setLoading] = useState(true);
-
   useEffect(()=>{
     setTimeout(()=>{
       setLoading(false);
@@ -19,23 +19,25 @@ function App() {
   },[])
 
   return (
-    <div>
-        {loading ? <Loader /> : 
-          <div>
-            <BrowserRouter>
-              <Routes>
-                <Route path="/" element={< Layout/>}>
-                  <Route index element={<Home/>} />
-                  <Route path="about" element={<AboutUsPage/>} />
-                  <Route path="why-pocketme" element={<WhyPocketmePage/>} />
-                  <Route path="how-it-works" element={<HowItWorks/>} />
-                  <Route path="contact-us" element={<ContactUs/>} />
-                </Route>
-              </Routes>
-            </BrowserRouter>
-          </div>
-        }
-    </div>
+    <RecoilRoot>
+      <div>
+          {loading ? <Loader /> : 
+            <div>
+              <BrowserRouter>
+                <Routes>
+                  <Route path="/" element={< Layout/>}>
+                    <Route index element={<Home/>} />
+                    <Route path="about" element={<AboutUsPage/>} />
+                    <Route path="why-pocketme" element={<WhyPocketmePage/>} />
+                    <Route path="how-it-works" element={<HowItWorks/>} />
+                    <Route path="contact-us" element={<ContactUs/>} />
+                  </Route>
+                </Routes>
+              </BrowserRouter>
+            </div>
+          }
+      </div>
+    </RecoilRoot>
   )
 }
